move consolidator tests to test/ dir and run them via 'npm test'
